requirements.txt
gunicorn
django-heroku
plotly